[metrics64] Implement generic tag schemas
I'm imagining a future where we can derive a metric definition from a struct.
To support that, along with the existing "just a bunch of strings" definitions,
I've added a bunch of trait fuckery to the MetricDefinition layer. From the
registry down we still only operate on &str and &'static str.