[metrics64] Initial exponential histogram implementation
Oh boy, we've got histograms! For this first pass I've pretty much copied the
upstream go otel implementation. I need time to fully understand how this thing
works, but for now the basics are ready. Its been piped into the registry and
we track export latency with it which is cool.

Two core outstanding bits that'll come later.
1. No tests, I need to port the test suite from go, and probably expand it
2. How the fuck do I export quantiles from this thing? They say its basically
   the same as prometheus, so I can probably go rip their quantile estimation
   code out.