Implement frexp in rust, fix precision issues in ldexp
I've copied the musl implementation for frexp. Then I did some fuzzing against
both the rust implementations to try and suss out any differences with libm.
It worked! There was a integer precision error in ldexp, and my initial port
of frexp had a bug with the 'truthiness' of a f64.

I'm leaving the debug_assertions gated validation logic in place for a while to
really make sure its accurate.